requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==4.2.4
django-crontab==0.7.1
Pillow==10.0.0
redis==5.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3

python-dateutil~=2.8.2